+++ Plugins/Race/Race.hpp
@@ -101,6 +101,9 @@ private:
     static void LoadRaceInfoHook(CNWRules*);
     static int32_t CheckItemRaceRestrictionsHook(CNWSCreature*, CNWSItem*);
     static int32_t GetFavoredEnemyBonusHook(CNWSCreatureStats*, CNWSCreature*);
+    static int32_t GetAttackModifierVersusHook(CNWSCreatureStats *pStats, CNWSCreature* pCreature);
+    static int16_t GetArmorClassVersusHook(CNWSCreatureStats* pStats, CNWSCreature* pCreature, BOOL bVsTouchAttack);
+
 };
 
 }